# backend/public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"TransactionCategory": {
				"enum": [
					"PAYMENT",
					"DEP",
					"CREDIT",
					"OTHER",
					"POS",
					"CASH"
				],
				"type": "string"
			},
			"Transaction": {
				"properties": {
					"category": {
						"type": "string"
					},
					"serial": {
						"type": "string"
					},
					"categories": {
						"$ref": "#/components/schemas/TransactionCategory"
					},
					"balance": {
						"type": "number",
						"format": "double"
					},
					"creditAmount": {
						"type": "number",
						"format": "double"
					},
					"debitAmount": {
						"type": "number",
						"format": "double"
					},
					"narrative": {
						"type": "string"
					},
					"date": {
						"type": "string"
					},
					"bankAccount": {
						"type": "string"
					},
					"id": {
						"type": "string"
					},
					"comment": {
						"type": "string"
					}
				},
				"required": [
					"serial",
					"categories",
					"balance",
					"creditAmount",
					"debitAmount",
					"narrative",
					"date",
					"bankAccount",
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TransactionType": {
				"properties": {
					"category": {
						"type": "string"
					},
					"serial": {
						"type": "string"
					},
					"categories": {
						"$ref": "#/components/schemas/TransactionCategory"
					},
					"balance": {
						"type": "number",
						"format": "double"
					},
					"creditAmount": {
						"type": "number",
						"format": "double"
					},
					"debitAmount": {
						"type": "number",
						"format": "double"
					},
					"narrative": {
						"type": "string"
					},
					"date": {
						"type": "string"
					},
					"bankAccount": {
						"type": "string"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"serial",
					"categories",
					"balance",
					"creditAmount",
					"debitAmount",
					"narrative",
					"date",
					"bankAccount",
					"id"
				],
				"type": "object"
			},
			"Category": {
				"description": "Category to classify transactions together.",
				"properties": {
					"transactions": {
						"items": {
							"$ref": "#/components/schemas/Transaction"
						},
						"type": "array"
					},
					"icon": {
						"type": "string"
					},
					"description": {
						"type": "string",
						"description": "the description columns"
					},
					"title": {
						"type": "string"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"title",
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ObjectLiteral": {
				"description": "Interface of the simple literal object with any string keys.",
				"properties": {},
				"type": "object",
				"additionalProperties": {}
			},
			"UpdateResult": {
				"description": "Result object returned by UpdateQueryBuilder execution.",
				"properties": {
					"raw": {
						"description": "Raw SQL result returned by executed query."
					},
					"affected": {
						"type": "number",
						"format": "double",
						"description": "Number of affected rows/documents\nNot all drivers support this"
					},
					"generatedMaps": {
						"items": {
							"$ref": "#/components/schemas/ObjectLiteral"
						},
						"type": "array",
						"description": "Generated values returned by a database.\nHas entity-like structure (not just column database name and values)."
					}
				},
				"required": [
					"raw",
					"generatedMaps"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TransactionPatchRequest": {
				"properties": {
					"category": {
						"type": "string"
					},
					"comment": {
						"type": "string"
					},
					"serial": {
						"type": "string"
					},
					"narrative": {
						"type": "string"
					},
					"debitAmount": {
						"type": "string"
					},
					"date": {
						"type": "string"
					},
					"creditAmount": {
						"type": "string"
					},
					"categories": {
						"type": "string"
					},
					"bankAccount": {
						"type": "string"
					},
					"balance": {
						"type": "string"
					}
				},
				"type": "object"
			},
			"User": {
				"properties": {
					"password": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"id": {
						"type": "string"
					},
					"salt": {
						"type": "string"
					},
					"mobile": {
						"type": "string"
					},
					"basiqId": {
						"type": "string"
					}
				},
				"required": [
					"password",
					"email",
					"username",
					"id"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "budget-backend",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/api/categories": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Category"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Category"
				],
				"security": [],
				"parameters": []
			}
		},
		"/api/transactions": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Transaction"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Transaction"
				],
				"security": [],
				"parameters": []
			}
		},
		"/api/transactions/{id}": {
			"get": {
				"operationId": "GetById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Transaction"
								}
							}
						}
					}
				},
				"tags": [
					"Transaction"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"patch": {
				"operationId": "Patch",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UpdateResult"
								}
							}
						}
					}
				},
				"tags": [
					"Transaction"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TransactionPatchRequest"
							}
						}
					}
				}
			}
		},
		"/api/users": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/User"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/User"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/User"
							}
						}
					}
				}
			},
			"patch": {
				"operationId": "PatchUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UpdateResult"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/User"
							}
						}
					}
				}
			}
		},
		"/api/users/{id}": {
			"get": {
				"operationId": "GetById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/User"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"204": {
						"description": ""
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}